配置常量，统一放到 `io.seata.common.ConfigurationKeys` 中
