+++ config/seata-config-core/src/main/java/io/seata/config/ConfigurationKeys.java
@@ -19,30 +19,8 @@ package io.seata.config;
  * The type Configuration keys.
  *
  * @author slievrly
+ * @deprecated The constants are moved to {@link io.seata.common.ConfigurationKeys}
  */
-public interface ConfigurationKeys {
-    /**
-     * The constant FILE_ROOT_REGISTRY.
-     */
-    String FILE_ROOT_REGISTRY = "registry";
-    /**
-     * The constant FILE_ROOT_CONFIG.
-     */
-    String FILE_ROOT_CONFIG = "config";
-    /**
-     * The constant SEATA_FILE_ROOT_CONFIG
-     */
-    String SEATA_FILE_ROOT_CONFIG = "seata";
-    /**
-     * The constant FILE_CONFIG_SPLIT_CHAR.
-     */
-    String FILE_CONFIG_SPLIT_CHAR = ".";
-    /**
-     * The constant FILE_ROOT_TYPE.
-     */
-    String FILE_ROOT_TYPE = "type";
-    /**
-     * The constant DATA_TYPE.
-     */
-    String DATA_TYPE = "dataType";
+@Deprecated
+public interface ConfigurationKeys extends io.seata.common.ConfigurationKeys {
 }
