+++ sqlparser/seata-sqlparser-core/pom.xml
@@ -30,12 +30,5 @@
             <artifactId>seata-config-core</artifactId>
             <version>${project.version}</version>
         </dependency>
-
-        <dependency>
-            <groupId>${project.groupId}</groupId>
-            <artifactId>seata-core</artifactId>
-            <version>${project.version}</version>
-            <optional>true</optional>
-        </dependency>
     </dependencies>
 </project>
